Move response objects into appropriate submodules
