+++ pylastfm/response/library.py
@@ -1,0 +1,10 @@
+from figgis import Field
+from pylastfm.response.common import ApiConfig, _ArtistBase
+
+
+class Artist(ApiConfig):
+
+    __inherits__ = [_ArtistBase]
+
+    playcount = Field(int, required=True)
+    tagcount = Field(int, required=True)
